src/main/java/guru/springframework/recipe/services/RecipeServiceImpl.java
package guru.springframework.recipe.services;

import java.util.HashSet;
import java.util.Set;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import guru.springframework.recipe.domain.Recipe;
import guru.springframework.recipe.repositories.RecipeRepository;
import lombok.extern.slf4j.Slf4j;

@Slf4j
@Service
public class RecipeServiceImpl implements RecipeService {
	
	private final RecipeRepository recipeRepository;

	@Autowired
	public RecipeServiceImpl(RecipeRepository recipeRepository) {
		super();
		this.recipeRepository = recipeRepository;
	}

	@Override
	public Set<Recipe> getRecipes() {
		log.debug("Getting recipes...");
		Set<Recipe> recipes = new HashSet<Recipe>();
		this.recipeRepository.findAll().iterator().forEachRemaining(recipes::add);
		
		return recipes;
	}
	
}
